Router Functionality
Moved routes from app.js to routes/ folder and individual script files
