+++ views/partials/footer.html
@@ -1,6 +1,6 @@
     
     <!-- Footer -->
-    <footer class="footer" style="background-image: url('{{ footer.metadata.footer_bg.imgix_url }}?w=1000&fit=crop&crop=entropy&auto=format,enhance')">
+    <footer class="footer" style="background-image: url('{{ cosmic.object.footer.metadata.footer_bg.imgix_url }}?w=1000&fit=crop&crop=entropy&auto=format,enhance')">
     <div class="h-100 overlay" style="background:rgba(0,0,0,0.6);">
         <div class="text-center">
             <img class="img-fluid logo" src="{{ logos.metadata.text_logo.url }}" alt="">
