+++ views/partials/header.html
@@ -77,7 +77,7 @@
     <nav class="navbar fixed-top navbar-toggleable-md navbar-inverse navbar-expanded" id="mainNav">
         <div class="container" style="display:flex;">
             <a class="navbar-brand page-scroll img-fluid" href="/">
-              <img src="{{ logos.metadata.text_logo.url }}" height="120" width="450" alt="Sloma">
+              <img src="{{ cosmic.object.logos.metadata.text_logo.url }}" height="120" width="450" alt="Sloma">
             </a>
             <div class="navbar-collapse collapse justify-content-end" id="navbarResponsive">
                 <ul class="navbar-nav">
